pyproject.toml
[project]
name = "embodied-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kapture>=1.1.10",
    "numpy>=2.3.4",
    "openai>=2.8.0",
    "opencv-python>=4.11.0.86",
    "rerun-sdk>=0.27.2",
    "sentence-transformers>=5.1.2",
]
